Python/requirements.txt
fastap==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.37
pydantic==2.10.5
requests==2.28.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
argparse==1.4.0